+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [keep a changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+## [0.1.1] - 2017-10-18
+- Compatible with Swift 4.0
+- Compatible with Nakama 1.1.0
+- Added Carthage and Cocoapods support.
+
 ## [0.1.0] - 2017-08-22
 ### Added
 - Initial public release.